requirements.txt
python-dotenv==1.0.1
langchain-google-genai>=1.0.5
langchain-community>=0.2.10
langchain-chroma==0.2.5
langchain==0.2.12
langchain-core>=0.2.12
pypdf==4.3.1
bs4==0.0.2
chromadb==0.6.3
